procstream/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.8.6
    # via
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via langchain
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
beautifulsoup4==4.12.2
    # via
    #   feedfinder2
    #   newspaper3k
cassandra-driver==3.28.0
    # via cassio
cassio==0.1.3
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   pulsar-client
    #   requests
charset-normalizer==3.3.1
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via
    #   geomet
    #   nltk
cssselect==1.2.0
    # via newspaper3k
dataclasses-json==0.6.1
    # via langchain
exceptiongroup==1.1.3
    # via anyio
feedfinder2==0.0.4
    # via newspaper3k
feedparser==6.0.10
    # via newspaper3k
filelock==3.12.4
    # via tldextract
fire==0.5.0
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
geomet==0.2.1.post1
    # via cassandra-driver
idna==3.4
    # via
    #   anyio
    #   requests
    #   tldextract
    #   yarl
jieba3k==0.35.1
    # via newspaper3k
joblib==1.3.2
    # via nltk
jsonpatch==1.33
    # via langchain
jsonpointer==2.4
    # via jsonpatch
langchain==0.0.322
    # via -r requirements.in
langsmith==0.0.52
    # via langchain
loguru==0.7.2
    # via -r requirements.in
lxml==4.9.3
    # via newspaper3k
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
newspaper3k==0.2.8
    # via -r requirements.in
nltk==3.8.1
    # via newspaper3k
numpy==1.26.1
    # via
    #   cassio
    #   langchain
openai==0.28.1
    # via -r requirements.in
orjson==3.9.9
    # via -r requirements.in
packaging==23.2
    # via marshmallow
pillow==10.1.0
    # via newspaper3k
pulsar-client==3.3.0
    # via -r requirements.in
pydantic==2.4.2
    # via
    #   langchain
    #   langsmith
pydantic-core==2.10.1
    # via pydantic
python-dateutil==2.8.2
    # via newspaper3k
pyyaml==6.0.1
    # via
    #   langchain
    #   newspaper3k
regex==2023.10.3
    # via nltk
requests==2.31.0
    # via
    #   cassio
    #   feedfinder2
    #   langchain
    #   langsmith
    #   newspaper3k
    #   openai
    #   requests-file
    #   tldextract
requests-file==1.5.1
    # via tldextract
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   cassandra-driver
    #   feedfinder2
    #   fire
    #   geomet
    #   python-dateutil
    #   requests-file
sniffio==1.3.0
    # via anyio
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.22
    # via langchain
tenacity==8.2.3
    # via langchain
termcolor==2.3.0
    # via fire
tinysegmenter==0.3
    # via newspaper3k
tldextract==5.0.1
    # via newspaper3k
tqdm==4.66.1
    # via
    #   nltk
    #   openai
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.0.7
    # via requests
yarl==1.9.2
    # via aiohttp
